Clean up ContentAddressedStorage wrappers with DelegatingStorage
