+++ src/peergos/shared/storage/DelegatingStorage.java
@@ -1,0 +1,110 @@
+package peergos.shared.storage;
+
+import peergos.shared.cbor.*;
+import peergos.shared.crypto.hash.*;
+import peergos.shared.io.ipfs.multihash.*;
+import peergos.shared.user.fs.*;
+import peergos.shared.util.*;
+
+import java.util.*;
+import java.util.concurrent.*;
+
+public abstract class DelegatingStorage implements ContentAddressedStorage {
+
+    private final ContentAddressedStorage target;
+
+    public DelegatingStorage(ContentAddressedStorage target) {
+        this.target = target;
+    }
+
+    @Override
+    public CompletableFuture<BlockStoreProperties> blockStoreProperties() {
+        return target.blockStoreProperties();
+    }
+    @Override
+    public CompletableFuture<Multihash> id() {
+        return target.id();
+    }
+
+    @Override
+    public CompletableFuture<TransactionId> startTransaction(PublicKeyHash owner) {
+        return target.startTransaction(owner);
+    }
+
+    @Override
+    public CompletableFuture<Boolean> closeTransaction(PublicKeyHash owner, TransactionId tid) {
+        return target.closeTransaction(owner, tid);
+    }
+
+    @Override
+    public CompletableFuture<List<Multihash>> put(PublicKeyHash owner, PublicKeyHash writer, List<byte[]> signatures, List<byte[]> blocks, TransactionId tid) {
+        return target.put(owner, writer, signatures, blocks, tid);
+    }
+
+    @Override
+    public CompletableFuture<Optional<CborObject>> get(Multihash hash) {
+        return target.get(hash);
+    }
+
+    @Override
+    public CompletableFuture<List<Multihash>> putRaw(PublicKeyHash owner, PublicKeyHash writer, List<byte[]> signatures, List<byte[]> blocks, TransactionId tid) {
+        return target.putRaw(owner, writer, signatures, blocks, tid);
+    }
+
+    @Override
+    public CompletableFuture<Optional<byte[]>> getRaw(Multihash hash) {
+        return target.getRaw(hash);
+    }
+
+    @Override
+    public CompletableFuture<List<Multihash>> pinUpdate(PublicKeyHash owner, Multihash existing, Multihash updated) {
+        return target.pinUpdate(owner, existing, updated);
+    }
+
+    @Override
+    public CompletableFuture<List<Multihash>> recursivePin(PublicKeyHash owner, Multihash hash) {
+        return target.recursivePin(owner, hash);
+    }
+
+    @Override
+    public CompletableFuture<List<Multihash>> recursiveUnpin(PublicKeyHash owner, Multihash hash) {
+        return target.recursiveUnpin(owner, hash);
+    }
+
+    @Override
+    public CompletableFuture<Boolean> gc() {
+        return target.gc();
+    }
+
+    @Override
+    public CompletableFuture<List<Multihash>> getLinks(Multihash root) {
+        return target.getLinks(root);
+    }
+
+    @Override
+    public CompletableFuture<Optional<Integer>> getSize(Multihash block) {
+        return target.getSize(block);
+    }
+
+    @Override
+    public CompletableFuture<List<FragmentWithHash>> downloadFragments(List<Multihash> hashes,
+                                                                       ProgressConsumer<Long> monitor,
+                                                                       double spaceIncreaseFactor) {
+        return target.downloadFragments(hashes, monitor, spaceIncreaseFactor);
+    }
+
+    @Override
+    public CompletableFuture<List<PresignedUrl>> authReads(List<Multihash> blocks) {
+        return target.authReads(blocks);
+    }
+
+    @Override
+    public CompletableFuture<List<PresignedUrl>> authWrites(PublicKeyHash owner,
+                                                            PublicKeyHash writer,
+                                                            List<byte[]> signedHashes,
+                                                            List<Integer> blockSizes,
+                                                            boolean isRaw,
+                                                            TransactionId tid) {
+        return target.authWrites(owner, writer, signedHashes, blockSizes, isRaw, tid);
+    }
+}
