+++ src/peergos/shared/storage/HashVerifyingStorage.java
@@ -4,7 +4,6 @@ import peergos.shared.cbor.*;
 import peergos.shared.crypto.hash.*;
 import peergos.shared.io.ipfs.cid.*;
 import peergos.shared.io.ipfs.multihash.*;
-import peergos.shared.user.fs.*;
 import peergos.shared.util.*;
 
 import java.util.*;
@@ -12,12 +11,13 @@ import java.util.concurrent.*;
 import java.util.function.*;
 import java.util.stream.*;
 
-public class HashVerifyingStorage implements ContentAddressedStorage {
+public class HashVerifyingStorage extends DelegatingStorage {
 
     private final ContentAddressedStorage source;
     private final Hasher hasher;
 
     public HashVerifyingStorage(ContentAddressedStorage source, Hasher hasher) {
+        super(source);
         this.source = source;
         this.hasher = hasher;
     }
@@ -42,48 +42,6 @@ public class HashVerifyingStorage implements ContentAddressedStorage {
                 throw new IllegalStateException("Incorrect identity hash! This shouldn't ever  happen.");
             default: throw new IllegalStateException("Unimplemented hash algorithm: " + claimed.type);
         }
-    }
-
-    @Override
-    public CompletableFuture<Multihash> id() {
-        return source.id();
-    }
-
-    @Override
-    public CompletableFuture<List<PresignedUrl>> authReads(List<Multihash> blocks) {
-        return source.authReads(blocks);
-    }
-
-    @Override
-    public CompletableFuture<List<FragmentWithHash>> downloadFragments(List<Multihash> hashes,
-                                                                       ProgressConsumer<Long> monitor,
-                                                                       double spaceIncreaseFactor) {
-        return source.downloadFragments(hashes, monitor, spaceIncreaseFactor);
-    }
-
-    @Override
-    public CompletableFuture<List<PresignedUrl>> authWrites(PublicKeyHash owner,
-                                                            PublicKeyHash writer,
-                                                            List<byte[]> signedHashes,
-                                                            List<Integer> blockSizes,
-                                                            boolean isRaw,
-                                                            TransactionId tid) {
-        return source.authWrites(owner, writer, signedHashes, blockSizes, isRaw, tid);
-    }
-
-    @Override
-    public CompletableFuture<TransactionId> startTransaction(PublicKeyHash owner) {
-        return source.startTransaction(owner);
-    }
-
-    @Override
-    public CompletableFuture<Boolean> closeTransaction(PublicKeyHash owner, TransactionId tid) {
-        return source.closeTransaction(owner, tid);
-    }
-
-    @Override
-    public CompletableFuture<Boolean> gc() {
-        return source.gc();
     }
 
     @Override
@@ -124,30 +82,5 @@ public class HashVerifyingStorage implements ContentAddressedStorage {
                 .thenCompose(arrOpt -> arrOpt.map(bytes -> verify(bytes, hash, () -> bytes)
                         .thenApply(Optional::of))
                         .orElseGet(() -> Futures.of(Optional.empty())));
-    }
-
-    @Override
-    public CompletableFuture<List<Multihash>> pinUpdate(PublicKeyHash owner, Multihash existing, Multihash updated) {
-        return source.pinUpdate(owner, existing, updated);
-    }
-
-    @Override
-    public CompletableFuture<List<Multihash>> recursivePin(PublicKeyHash owner, Multihash h) {
-        return source.recursivePin(owner, h);
-    }
-
-    @Override
-    public CompletableFuture<List<Multihash>> recursiveUnpin(PublicKeyHash owner, Multihash h) {
-        return source.recursiveUnpin(owner, h);
-    }
-
-    @Override
-    public CompletableFuture<List<Multihash>> getLinks(Multihash root) {
-        return source.getLinks(root);
-    }
-
-    @Override
-    public CompletableFuture<Optional<Integer>> getSize(Multihash block) {
-        return source.getSize(block);
     }
 }
